# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant API",
    "version": "1.0.0"
  },
  "servers": [{"url": "http://localhost:3000"}],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login to get token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {"type": "string"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Successful login"}
        }
      }
    },
    "/api/menu": {
      "get": {
        "summary": "Get all menu items",
        "security": [{"bearerAuth": []}],
        "responses": { "200": {"description": "List of menu items"} }
      }
    },
    "/api/tables": {
      "get": {
        "summary": "Get all tables",
        "security": [{"bearerAuth": []}],
        "responses": { "200": {"description": "List of tables"} }
      }
    },
    "/api/tables/{id}": {
      "put": {
        "summary": "Update table status",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {"type": "string"},
                  "current_order_id": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": { "200": {"description": "Success"} }
      }
    },
    "/api/orders": {
      "get": {
        "summary": "Get all orders",
        "security": [{"bearerAuth": []}],
        "responses": { "200": {"description": "List of orders"} }
      },
      "post": {
        "summary": "Create a new order",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {"type": "string"},
                  "tableNumber": {"type": "integer"},
                  "totalAmount": {"type": "number"},
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {"type": "integer"},
                        "quantity": {"type": "integer"}
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": { "201": {"description": "Order created"} }
      }
    },
    "/api/orders/{id}": {
      "put": {
        "summary": "Update an order",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {"type": "string"},
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {"type": "integer"},
                        "preparationProgress": {"type": "integer"}
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": { "200": {"description": "Order updated"} }
      }
    }
  }
}